requirements.txt
PyQt6>=6.5
Pillow>=10.0
PyYAML>=6.0
anthropic>=0.30
psutil>=5.9
